1.37x diff2-kernel / 1.75x wavg-kernel
This from using the __ldg() intrinsic for some of the global loads;
it routes the load through the texture cache, meaning one may score
some hits. NOTE: Not all global loads benefitted from __ldg(), and
were omitted from this commit, BUT THE DEPENDENCY OF THE PERFORMANCE
INCREASE WAS NOT TESTED ON OTHER HARDWARE. The use of __ldg() herein
may thus not be the optimal across other hardware.

ALSO:  __ldg() requires compute sm_35 or higher, thus the flag was
added to the nvcc build flags. Fallback should perhaps be made.