GPU Reduce_wdiff2s-kernel 1.0
This makes for a ~50x-70x speedup of the image reduction, which
before took ~50ms on my hardware (Profiling build). However, it
could be improved by increasing the number of image-loads per
block, as this would decrease loads and writes per compute.

The kernel-call should also be able to handle more blocks than
the maximum 65535 in a single block-grid dimension, which would
be partially fixed by the above-mentioned improvement. Otherwise
a simple multidim-blockgrid solution is fine.