Sumweight kenrel does both coarse and fine sampled
passes. The previous kernel used 32 threads to access
oversampled points, which are adjacent in global mem
and therefore very fast to load. In order for the kernel
to generalize for coarse points, the access is now made
with each thread fetching an (over)sampled orientation,
which are in multiples of 32 (so as to keep all threads
occupied), but not adjacent in global memory. The mem-
ory access should therefore be far less efficient, but
the code works equally well for all healpix- and over-
sampling, provided healpix_order>1.