PAV_TTE and PAV_TGE fixed, double build works
They had two separate issues;

1 - One input int type was unsigned instead of long,
which became an issue because it actually needed to
be signed. thus -33 -> HUGE NUMBER

2 - The padding_factor amendments were only imple-
mented on the PAV_TTI and PAV-combo kernels. This
meant that when double precision forced a split of
PAV and WAVG, it forced usage of PAV_TGE, which was
unaware that it should correct the euler matrix by
padding factor.