Indexing issue in higher order coarse pass fixed
Previous code relied on the assumption that all coarse orientations
should be difference-calculated, regardless of being the first or any
later refinement-iteration. This is apparently not true, as evident by
the if-statement in generateProjectionSetup which is dependent on
pdf_orientation being finite.

The subsequent errors with/without --ctf flag and for fewer particles,
is then likely a result of the varying probability that there should
be some overlap between the orientational priors (weights) and the
erraneously mapped differences. If it happened to be zero overlap,
the result would be sumweights=0 and exit=failure, otherwise the
algorithm would continue and exit with success, but have completely
nonsensical output data, since weights and priors are then utterly
mismatched.