Partial removal of dependence on sparse Mweight
Current code relies on the implicit indexing of diffs and
weights by mapping the into an array "Mweights", which in
the gpu-based code belongs to the op class. For somewhat
converged data this tends to be a VERY sparesly populated
array by way of the low significance of many of the asso-
ciated orientations of an image. Combined with the large
size of the array with fine sampling, the cost of

a) transferring this object to and from the gpu
b) keeping a largely empty array on the gpu and
   cycling through comparisons of insignificant
   elements

becomes too large compared to the actual computations able
to complete. The solution is also twofold;

A) move the object up in scope and keep it on the gpu
B) coalesce data my keeping an explicit indexing, rather
   than implicit mapping

This commits adresses both for the fine pass of difference-
calculation, while still mapping to the Mweight array at
the end of converting differences to weights. This because
the storeWeightedSums()-function has not yet been adapted
to handle explicit indexing rather than implicit mapping.