Getrow & nonZero-picking removed from fine pass
These apparently only constitute a significant part
of the CPU-bound code in the fine pass. That taken
together with the fact that the Mweight array is
used for much additional implicitly indexed checks
(like setting coarse significance for the fine pass)
makes it easier to (at the moment) let the two methods
(sparse Mweight and dense weights) coexist and simply
use the latter in the fine pass where it removes the
need to fetch a large section of an even larger
Mweight-array isolate non-zero weights.