ArrayMasks for classes (multi-class capab. prep)
We use dense arrays with explicit indexes for the sparse pass,
where oversampling could make the very few significnatly con-
tributing expensive to iterate through with nothing to do. At
some stages, we need to perform reductions of all classes. This
offers a solution;

Make an explicitly indexed array, carrying weights and indices
for all classes of this image. That way any class-wide reduction
has access to coalesced data with direct indexing. When we need
data for a specific class, we use a mask which when fed into a
constructor for a sub-indexedArray, returns a handle for the data
belonging to that class. The mask also holds arrays which deter-
mine the job-partion necessary for some kernels to exploit re-
use of some data.

The "global" index array for all classes of an image is (in
some sense) appended by the data of each new class as signficance
of each finely sampled gridpoint is determined, and the mask of
each class is simultaneoulsy made, to be available in the conv-
ersion of weights and the backprojection of them. The job parti-
tion of the mask is independent of the indexedArray, and can be
re-defined for the purposes of any kernel designed to use them.