2 classes pass second iter, less errs in metadata
A shift in usage compared to previous commits
to enable multi-class refinement;

before:
only the classes which had significant orinetations to be examined were
prepared, by appending any new classes and weights to existing objects.
The result of this was a difficult-to manage class indexing, since loops
are over all classes, rather than all classes which are implicity deemed
relevant.

after:
Make space for all classes in class-arranging masks and ProjectionParam-
arrays, while keeping the weights-array coalesced. This way all classes
can be accessed by explicit indices in the masks and ProjParams-arrays,
and simply skipped when nothing is deemed significant in this class

to fix:
Some stuff is being done without purpose; breaks and continues should be
injected where nothing is to be done, so as make the run proceed quicker .
Some general cleanup might be good as well, and the error with increasing
particle number persists.