2 classes pass second iter for many particles
This commit fixes an error not apparent for low numbers of particles,
where the first iteration produces probability densities which are so
sparse that the multi-class passes essentially only ever tries the
same class which they were randomly assigned in the first iteration.

The error was that the weights which were determined for several classes
into a coalesced array, were never actually split in storeWS, but acc-
identally just used in its entirety. As long as the sparse probability
density limited the number of classes for all particles to a single one,
this was exactly equivalent to what was inteded. With more particles,
not splitting the array with weights for all classes caused undefined
behaviour, and eventually segfaults.

fixes #28