Template difference kernels based on 2D/3D-proj
This will keep a single diffCoarse-kernel in the source-code, but
expand the compiled code to include separate kernels for 2D- and
3D-projections, respectively. This to avoid an if-statement being
evaluated once for every pixel treated in the difference-kernels.

The above is also implemented for the diffFine-kernel.

There is probably some additional streamlining of the source code
to be done, for instance the calls to the templated functions are
runtime-dependent and stated explicitly for either case, both out
of necessity (the comiler should make both) and simplicity. It's
not too pretty though ( see cuda_helper_functions.cu:518:539, for
instance )