Reorder iclass > ipart loop in storeWS, step 2
This commit does 2 fixes in one action. The action is
to switch the index hierarchy of the first storeWS loop-
block as follows;

-- old version --
for iclass
    for proj_div
	for ipart

-- new version --
for ipart
    for iclass
	for proj_div

the fixes are:

1. The old hierarcy implicitly assumes that the same
projdivision would work for all iparts, which is not
true, thus it could never have worked for more than
one ipart; it needs to be the lowest index

2. Allows the completion of all classes for each ipart,
which is preferrable since it is building an indexed
Array with coalesced access over all classes for each
ipart. To avoid injections and reallocations of alot of
data for each ipart, this is preferrable.

Also: One associated if(do_x && do_y)-block is replaced
by a if(!do_x || !do_y)-continue case.