Vectorization of Fine-pass weight objects
This commit represents the capability of the GPU-
based algorithm to handle sub-images of what in the
code is called and ori_particle, which is essen-
tially a particle which was picked in a cryo-EM
micrograph. Collecting movies instead om images
produces precisely such sub-images of an obesrved
particle, and thus this commit add movie-capability
to the current GPU-code.

NOTE:   THIS IS AN UNTESTED FEATURE, SUPPORT IS
CURRENTLY THEORETICAL, AS ONLY SINGLE-IMAGE PART-
ICLE TEST HAVE BEEN DONE.

untested fix #38