Forcing 2D runs to abide to psi-sammplings mod 32
Some kernels in the convert-stage of the coarse pass currently
relies on the number of orietnations to be divisible by 32, since
this is an efficient thread-wise division of work. During 2D-
classification the diretional sampling provided by the Healpix
mesh is not factored in, yielding very low orientational numbers,
the default being every 10degrees => 36 sampling points. This will
unfortunately not be compatible with current kernels.

This commit fixes this in the simplest and fastest way possible,
by forcing psi_step to increments which produce 0 mod 32 orienation
numbers. The default is then implicitly modified as

10degrees => 36 sampling points , ROUND(/32) -> 32 sampling points,
so psi_step is modified to 360 / 32 = 11.25

NOTE: this commit has issus with texture interpolation, so to see this
edit targeting issue #40, double precision build is required.