2D classification proj uses textures correctly
The layout for object texture creation follows very different methods,
opposite to remarks made in commit ec377f9. The hope was that the
existing method would be able to create a 2D-texture capable of using
bilinear interpolation, as per the CUDA documentation. This commit
resorts to the 'default' method of allocating a pitched (aligned) cuda-
array and uses bilinear interpolation intrinsics during 2D classfication.

fixes issue #40