Index fix in 2D priors & psi-domain control
An array reduction was wrongfully using the iorient index to read from a
reference an array, whereas it should be indexed by a pure iteration var.
The reulting reduction was used as prior in iterations>=2 during 2Dclass-
ification, so no 3D test up until now have been sensitive to this error.

Also, the first iteration produced correct meta-data, apart from the psi-
angle which was set to the [-180,180[ domain for the CPU-only code, where
the GPU code gave output in th [0,360[ domain. The reason is not fully
clear but an explicit control has been added regardless of using GPU hard
ware or not.

NOTE : the psi-angle domain control should be tested on iteration >2 when
possble, to rule out any possible issues with the extra domain check in
both 2D- and 3D-classification.

this commit resolves ticket #43, but there are further errors persiting,
which are described in ticket #46