2D classif. runs 2 iter w/o metadata errors
In order to swiftly implement 2D BackProjection, this was placed on the CPU
hardware. Due to this, the copying back of an unused device BackProjector
caused the CPU-Backprojector to be overwritten, resulting in an essentially
NULL comparison between images, since they were compared to all-zero models.
The only difference between weights was then due to the error associated
with interpolating pixel-values of the rotated images.

NOTE: there are still accuracy errors on single precision, the "w/o errors"
declared above refers to the obvious erros in the method implementation.

fixes #46