Moved two thrust-wrapped reducts out of class-loop
It was noticed that there was a large overhead associated with
launching thrust-kernels. In the 3D-case this overhead is negligble
compared to the time it takes to run the reductions, however, in
the 2D case the problem size is so much reduced that theis is no
longer the case. This commit represent edits to the code which
benefits the 2D-classification by grouping reductions into a single
call over all classes. It will benefit the 3D-classification to
the same absolute extent, however that might not even be noticable.

The changes are made to thrust reductions
 - min(diff) @ end of diff2coarse
 - max_element(weight) @ end of storeWS