new class cudaStager to coalesce _small_ mem-ops
Given a number of cudaGlobalPtr:s wit hthe same type, they can be
handed off to a stager which will move their memory positions into
a coalesced memory space and transfer them simultaneously.

A number of improvements are possible:

- optional use of pinned memory as staging space
this relies on whiching to management of a copy of the supplied
cudaGLobalPointer, rather than reassigning the only existing copy

- allocate space only once total size is known
this relies on the development of an appendable list
of cudaGlobalPtr:s to reassign host- and device-pointers of, at stage-
time.