Added missing loop for norm-corr factor
This was somehow glanced over and manifested as a small error which
could be attributable to larger particle groups being affected by a
different order of summation on the GPU comapred to the CPU.

NOTE 1:
The errors in METADATA which this commit adresses only appear when
multiple micrographs are used.

NOTE 2:
There is still some (reduced) deviation from CPU normcorr-values
when using double prec. GPU code. This could be additional factors
which are unaccounted for.

NOTE 3:
Using a .star-file which conatins 100s of particles from a single
micrograph and a single particle from another micrograph, the error
of the norm-correction of the single particle compared to CPU-code
is much higher than that of other particles. This (possible issue
is entered as issue #54

Fixed issue #53