Added missing is-statement in partial class-loop
The major class-loop in SWS has been partitioned into pre-process,
memory transfer, kernel-calls, and post process loops. In the final
such loop, an if-statement was missing which may have caused some
segfault under certain circumstances. It is unclear wheter this is
the fix to, but it was certainly found because of, an error similar
to that described in issue #57