Cross-correlation works (CC)
The kernels are NOT optimised for --always_cc, but rather just
to be compatible with the need for --firstiter_cc, whereunder we
simply assume that a stupid implementation will not affect the over-
all perfomrnace of an entire classification/refinement run.

NOTE: A third factor in the GPU-based image-correction applied to
the difference metric is introduced, which should cause some concern;
the accuracy might in some cases suffer. This commit leaves an
alternate availiable; however commented out. op.local_sqrtXi2 can be
applied either within the buildCorrImage, or at the end of CC-based
difference kernels. The present approach is the former.

fixes issue #11, creates optimisation issue #58