SoftMaskOutside on GPU
until tested by time, a bool remains to switch on native (CPU) function.

ALOT of optimization to be done, at the moment a ~2-3x speedup of
SoftMaskOutside-block and wrapping ZeroMask-block.

This could be reduced by half; A kenrel which prepares a [0,1]-domain
filter for all particles could devote SoftMaskOutside() to figuring out
just the scaling and apply a pix-by-pixel filter without the need to
check all norms and range-conditionals within the kernel.