SumW-kernels overlap, using cudaMLO class-streams
In order to keep the number of streams down, the classstreams
were made a member of the cudaMLO-class.

NOTE : at the moment we use DeviceSyncs after the only class loop
which uses this class-streaming. This might be bad, so we should
think about some way of syncing just these streams before continuing
along the CPU-code.