Updated warns and non-0 mod 32 compatibility
Previously thew orientational sampling was implicitly assumed to be
0 mod 32 in the the coasre pass due to the healpix sampling. When
doing 2D classification the psi_step dictates the number of increments,
the default being 10degrees => 36 steps. The quick and dirty fix was to
force 0 mod 32 by adjusting psi_step to fulfil psi_step = 360/(N*32)
for some N.

During local searches and/or auto-refine 32-multiple coarse psi-sampling
might not be fulfilled or even desirable for effciciency. To circumvent
this we have now allowed any psi-sampling, at the cost of wasted cycles
on the GPU. A cleaner fix from the point of view of functionality, but
still not the computationally sleakest fix.

ALSO some cleanup of unused class_ids in IndexedDataArray