cub appears to work, replaces thrust
The identified and amended issue was two-fold, and roughly independent.

The custom allocator was not correctly aligning memory addresses, resulting
in wrong and/or misaligned adressing.

Switching to the default allocator, the temporary storage size got set to
zero for small sorts, which resulted in a NULL temporary storage ptr, and
subsequent sort again tried to estimate the needed size, rather than
perform an actual sort.

this resolves issue #67