fftw is external project which fetches fftw-3.3.4
Looks for system libs, but builds local fftw-3.3.4 libs and headers
which are automatically included in the build, if system libs are
not found.

If built in the local directory, any new build will find these and
refrain from rebuilding the loacla fftw.

The proeccess is able to modify the CPU-precision through ccmake, although
new libraries will be built if they do not allready exist in the system or
local directories.