nvcc avoids float128 by ugly hack
gcc supports a type float128, which the wrapper nvcc does not support.
fftw has enabled support for this under the gcc compiler, but not for
the intel compiler. So in order to compile fftw using nvcc as a gcc-
wrapper, we can fool nvcc to think it is an intel compiler during fftw
compilation/inclusion.

The better solution would be for
  a) cuda to enable float128
  b) fftw to sense nvcc, rather than intel

Option bi) is possible through minor modification of the fftw3.h-header,
as described here:

https://github.com/FFTW/fftw3/issues/18

This has for some reason not made it into the past or current release
of ftw, however, so option a) is forced to provide compatibility or
RELION with past and present fftw-verisons without modification. This
is achieved by adding an nvcc-flag which provides an intel compiler
identifier. ( se ../cmake/BuildTypes.cmake)