Enabling a single complex texture is now possible
There is some doubt as to the benfit, however. One would expect that
3D projection kernels would benefit through decreased artimetic count
during texture-intrinsic interpolation. This is to some extent observed,
but the entire beenfit does not pervail due to a slowdown in other parts
of the code, for reasons unknown.

The projection-instructions are maintained such that a call separates
real and imaginary components. If (as in autopicker) the desired out-
put is a complex array, then these components are either joined after
projection or written separately.