GPU also shrinks the micrograph during calcs
NOTE:

This may be due to the difference in FFTW and CUFFT, but it may also
be the fact that the GPU has limited ammounts of memory. I think it
is the combinations them, but anyway:

Some micrograph sizes cause issues, which appear to be due to the
accidnetal or occational occurence of large prime factors in the
fourier-representation image sizes. LArge primes means alot of extra
(temp) storage during FFT calcs.

Why was this not an issue on the CPU?

a) CPU is serial -> onlye ever 1 FFT at a time
b) FFTW != CUFFT, algorithmically
c) y=1950 => x=976, so largest prime is 61. Big, but ok.

For, instance, trying y=1940 on the GPU with 5deg orient-step:

360/5 = 72 FFTs at once (in memory at once)
y=1940 => x=971 (which is prime), so largest prime is 971 (!)

So even though 1940 < 1950, the estimated space needed to perform
the FFT is more than 8 times bigger. combine that with performing 72
FFT at once, and the space requirement is a whopping 576 times larger.

This is not so much a problem as a hurdle, which is solved by two
considerations/modifications:

1 - Reduce 1950 much beyond 1940. The smaller the number, the smaller
the possibility of large prime, in addition to smaller objects and faster
calculations.

2 - Enforce numbers which doesn't contain large prime factors. GROMACS
allready does this, but a table lookup is not that complicated if we feel
like doing our own. The formal requirement is to force a basic dimension
x such that

  max(largest_prime_factor(x),largest_prime_factor((x-1)*2)) < M

where M is a cutoff which we specify as some ok-sized prime, like 61 or so.

FOR NOW, THIS IS AN ISSUE WITH THE GPU-CODE WHICH NEEDS TO BE FIXED