Scale particles picked from shrunken mic
Given that the calculations are performed on a shrunken micrograph,
this commit realized the idea of perfroming particle picking on this
image, then scaling coordinates to the original mircograph size. This
rather than making the image bigger simply to perform particle picking
on it.

The question remains, however; Does this affect the picking quality?

In test, shrinking affects the picked results at ~55% of original
dimesions, i.e. ~30% of original image size. This sounds reasonable,
but the resolution limit for this run (lowpass=20), is as low as ~35%
of image dimensions, i.e. ~13% of image size.

It should be noted that shrinking the micrograph only very slightly,
can improve the picking (in my subjective assesment, by vis. inspect.)