New syntax for device-affinity of ranks/threads
Syntax examples and explanation, assuming a machine with 4 GPUs;

A)
--gpu
No indices will distribute threads and ranks as uniformly as possible

B)
mpirun -n 3 ... -j 2 --gpu 0,2:1,3
A colon delimits each rank, and the integers denote the devices to be
used for that rank. In this case, the number of integers match the
number of threads so the mapping will be (slave-rank/thread):
(1/0)->0
(1/1)->2
(2/0)->1
(2/1)->3

C)
mpirun -n 4 ... -j 2 --gpu 0,2:1,3
As B), but the slave-rank which has no indices is treated as in A),
resulting in the same mapping as B), and additionally
(3/0)->2
(3/1)->1     (uniform distribution of this rank over all devices)

D)
mpirun -n 4 ... -j 3 --gpu 2:2:1,3
Slave ranks 1 and 2 are restricted to device 2, slave 3 is restricted to
devices 1 and 3. All ranks simply distribute themselves uniformly across
their restrictions, which results in a higher memory-load on device 2, and
no utilization of device 0 (note that each rank resident on a device has
shared memory object for its threads on that device, so device 2 has roughly
twice the memory load of device 1, in this case):
(1/0)->2
(1/1)->2
(1/2)->2
(2/0)->2
(2/1)->2
(2/2)->2
(3/0)->1
(3/1)->3
(3/2)->1