MPI-enabled autopicker uses GPUs
Same syntax as refinement, that is colon-separated indices for ranks.
As there is no thread-parallelism in the autopicker, comma-separated
indices are not necessary, and have no effect:

Default behaviour:
Requires NO indices to be suppied, and is device = rank % #devices,
so that if there are 3 devices and 5 threads, they will map as
(rank->device)
0->0
1->1
2->2
3->0
4->1

When mapping three ranks to three devices, threads (comma-sep) mean
nothing:
--gpu 0:1:2 <=> --gpu 0,0:1,2:2

To place two ranks on device 0 and one rank on device 1,
--gpu 0:0:1

Currently, the GPU-autopicker requires memory proportional to the
number of orientations (--ang 5 => 72 orientations), which can be
quite substantial. A single rank can handle the case where these
do not fit into the GPU memory, but ranks are NOT aware of each
other, meaning that multiple ranks may cause out-of-memory errors
in the present implementation. This will likely be fixed in the
future, so as to reduce memory usage and/or make ranks mutually
aware.