Loop thread-wise additive reduction across devices
Each thread adds it's contribution to the weights in a refinement-
iteration. Since threads on the same device share this object, the
weights get to be too large when threads in the same rank run on
the same device, since all add their joint contribution. Instead,
each such shared device-object should contribute.

this fixes issue #84

Note that issue #83 remains unfixed in this commit, such that the
extra normalization needs to be omitted in the GPU-code for a
direct CPU vs. GPU comparison to be valid.