system-safe memory artihmetic for GPU-allocs
The requested memory was formulatyed as

size_t mem = ... ;
int allocSize = (int) mem - ... ;
GPUmalloc(allocSize);

where the int-cast of "mem" was unsafe, in that if "mem" was larger
than the system int, a negative value was returned. The new method
uses size_t for all these cases, and where subtraction is needed,
the minuend and subtrahend (yes, that's what they're called, I
looked it up) are compared for size, not their difference against 0.

this fixes issue #85