Prep for mem-leak fixes
Purposely fragmented the mem-leak to allow future clean-up and/or
additional fixes to be smoother. On this commit a memleak is made
obvious under the default allocator. It was previously invisible
to cuda-memcheck, since there was no call to cudaDeviceReset().

A un unused function CudaGlobalPtr.resize() host is added, to be
used in a future commit.