Leak fix to default allocator.
The issue seems to be that the default template argument for any
CudaGlobalPtr is CuatomAlloc=True, regardless of whether the custom
allocator is used. This results in calls to cudaFree() being avoided.

It is unclear to me how the allocation can remain unaffected. None-
theless, the leak is fixed. Not only does cuda-memcheck report less
leaking memory, but the positive memory gradient (as reported by
nvidia-smi) is also reduced, in fact almost eliminated.