Leak fix to default allocator II
I the problem fixed here appears to be that the temporary array
allocated in some cub/thurst-dedicated functions get marked for
deletion without actually getting deleted, as only the custom
allocator collects the garbage merked for deletion. When the
default allocator is used, this makes for a leak of temporary
storage.

This in part adresses issue #88