Leak fix to default allocator III
The problem herein fixed, is that the projectors and backprojectors
in the MlDeviceBundle were not cleared before program termination.
Even worse, they appear to have been NEVER cleared, so that upon a
new iteration, new texture-arrays were simply malloc:Ed and copied
into the texture space. For this purpose, both back- and projectors
are now cleared in MlDeviceBundle::resetData(), just as is the
ProjectionPlans. This should be verified to be correct procedure.

Also, this was tested with non-MPI compatibility, so the MPI-version
should be updated to comply.