Indexing fix to pdf_orientations
When a particle was classified as contributing to more than
a single reconstructed class, the directions/orientations
were always selected starting from the beginning of the
array holding the direction-indices for all classes. This
mistake was made ONLY in the summation of the orientational
distribution, not the backprojection.

Essentially, each image contributed correctly to the
distribution of directions for the FIRST class it was class-
ified as, and contributed incorrectly to the distribution
of directions for all subsequent classes it was assigned to.

This means that class 0 always had the correct distribution
(up until and including 2nd iter), whereas other classes where
roughly correct. The latter because

a) many images were assigned as NOT contributing to class 0,
in which case they contributed correctly to the first class
they WERE assigned to.

b) the orientational assignment is roughly the same, irrespective
of class, at low convergence ( low class differentiation).

this fixes issue #95