Removed fftw_cleanup causing autopicker issue
The cleanup destroyed all references to fftw-objects, not just the
local transformer. Now we instead let it go out of scope and get
handled by the destructor, which cleans up the local copy only.

This fixes issue #99 (but we still don't know why that issue does
not appear when using a single reference as input for autopicking)