All objects in GPU-side autipicker are worksize
This affects the performance positively, as adjustments to FFT-sizes
based on prime-factors do not result in size-discrepancies. This also
allows us to remove an intermediate arrey for data resizing (FT-
windowing), resulting in a smaller memory footprint on the GPU.

In the present state, job segmentation is non-functional, but this
will be fixed in the next commit, which will restructure the psi-
batching in the GPU-code and clean up the autopicker code.