Adaptive OS used in fine
This should not be needed, but if oversampling is not used (i.e. if
there is only a single pass), then data-objects are incompatible with
storeWS in the GPU-version. We escape this by making a redundant calc
in a symbolic fine-pass that is NOT oversampled, but then we need
to use adaptive OS instead of hardcoding an oversampling in the fine
pass.