Backproject 3D template on 3D input
More instances of index checks that seem excessive, removed with
reference to the fact that we do fine without them in diff_coarse.
Prelim. tests (ctest no. 5 + subtomo_case) indicate this is ok,
but this should be furhter tested on much larger cases.

also: untested complementary changes to BP_2D