Auto-refine bugfix
This fixes a critical bug affecting half-set (gold-standard) refinement.

Starting with version 2.0.4, each iteration of auto-refinement incorrectly
uses the same half-set reconstruction as for alignment of all data. The FSC
and resolution reported is however still based on comparing two half-set
reconstructions which contain completely separate image data. A
degree of reference-bias is however present.  The effect is to report
a resolution higher than what should be under gold-standard conditions,
since, reconstructions are subject to a degree of over-fitting.

In order to claim a so-called "gold-standard" resolution, use any version
before and including 2.0.3, or after (including) this version.