Subset reconstruction handled sgd-like
Reconstruction should only occur when there has been something
back-projected. Using SGD, subsets might be small enough for
that not to happen, which results in bad estimations for the
cumulative weight. Code was written to deal with this, but it
did not cover the use of subsets under conventional (non-SGD)
execution. Runs which used subsets in conjunction with --maxsig
could therefore crash. This fix should avoid this issue.