tau2_fugde is NOT adjusted using non-SGD subsets
Mild confusion;

The issue that commit 10fd2e1 and d05ec6a was trying to solve was due
to the lack of an accuracy-check under the else()-case covering non-sgd
reconstruction. Those commits then removed the else-case to achieve this,
but also activated the tau2_fudge adjustment for non-SGD subset runs.

This commits reverts the latter effect, because tau2_fudge should ONLY
be subset-adjusted when using SGD, due to the update-formula used.