Register build byproducts for FFTW external.
If an external project is used for FFTW / FFTWF, Relion links libraries
directly against the resulting libraries (such as libfftw3f.so).

Some build generators (specifically, Ninja) require that the full dependency
graph is calculated at the beginning of the build process.  Since the
file `libfftw3f.so` does not exist at the beginning of the build (the
`ExternalProject` command has not been run yet) -- and there is no
rule to generate that file -- the build fails.

With this change, we introduce a fake target that relies on the
`ExternalProject` and has as a build byproduct the resulting library.
This allows `ninja` to correctly resolve the dependency and build Relion.

Signed-off-by: Takanori Nakane <nakane.t@gmail.com>