Refactor usage of `which`
In our environment it is not possible to use the `which` command,
hence we need to patch pipeline_jobs.cpp before we are able to compile
and use RELION.

Currently `which <cmd>` is used in several places in this file which
makes the patching a bit tedious. This commit adress this problem by
introducing a helper function obtainFullPathForCommand(), which converts
a string "cmd" into "`which cmd`". By using the helper function, it
means that we only need to change one row in this file before we are
able to build and use it.

Signed-off-by: Anton Haglund <aoh@nsc.liu.se>